update system for rooms + clean code
