+++ src/idrawable.ts
@@ -2,3 +2,7 @@
 export interface IDrawable {
 	draw(ctx: CanvasRenderingContext2D): void;
 }
+
+export interface IUpdatableDrawable extends IDrawable {
+	requires_update: boolean;
+}
